+++ panda/src/chan/animChannelMatrixDynamic.h
@@ -46,7 +46,7 @@ protected:
   AnimChannelMatrixDynamic();
 
 public:
-  virtual bool has_changed(int last_frame, int this_frame);
+  virtual bool has_changed(double last_frame, double this_frame);
   virtual void get_value(int frame, LMatrix4f &mat);
 
   virtual void get_value_no_scale_shear(int frame, LMatrix4f &value);
